+++ packages/patrol/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 - Fix `tapOnNotification` on iOS physical devices to Open button when tapping on notification. (#2972)
 - Fix `takeCameraPhoto` do not accept location permission dialog on Pixel physical devices. (#2972)
+- Fix `takeCameraPhoto` on Android to use fallback button if done button is not visible. (#2974)
 - Reflect failed tests in Playwright report. (#2970)
+
 
 ## 4.3.0